requirements.txt
ipython==8.27.0
ipykernel==6.29.5
openai==1.40.3
weaviate-client==4.7.1
sentence-transformers==3.0.1
tqdm==4.66.5
ipywidgets==8.1.5
ipywidgets==8.1.5
bokeh==3.5.2
pillow==10.4.0
